requirements.txt
nba_api>=1.4.1
pandas>=2.0.0
scikit-learn>=1.3.0
streamlit>=1.35.0
plotly>=5.18.0
pyyaml>=6.0
unidecode>=1.3.6
matplotlib>=3.7.0
